requirements.txt
# prod
aiohttp==3.11.14
pyjwt==2.10.1

# test
aioresponses==0.7.8

# dev
build==1.2.2
coverage==7.7.1
pip==25.0.1
ruff==0.11.2
